controller is done hopefully xD

+++ Backend/src/main/java/org/example/backend/MainForTesting.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class MainForTesting {
     public static void main(String[] args) {
-        int[][] graphMatrix = {
+        double[][] graphMatrix = {
                 {0, 10, 5, 0, 0},
                 {0, 0, 1, 25, 0},
                 {0, 0, 0, 6, 1},
